refactor(ios-home-tts): split Home tab TTS surface into focused files
- move `iOS/KeyVox iOS/Views/HomeTabView.swift` into `iOS/KeyVox iOS/Views/HomeTabView/HomeTabView.swift` so the Home tab can live as its own filesystem grouped feature area instead of a flat top level view file
- replace the monolithic `iOS/KeyVox iOS/Views/HomeTabView+TTS.swift` with a dedicated `iOS/KeyVox iOS/Views/HomeTabView/TTS/` folder so the Home TTS implementation is no longer mixed into one large extension file
- keep `iOS/KeyVox iOS/Views/HomeTabView/TTS/HomeTabView+TTS.swift` focused on the main Speak Copied Text card layout and composition only
- move transcript specific UI and visibility rules into `iOS/KeyVox iOS/Views/HomeTabView/TTS/HomeTabView+TTSTranscript.swift`, including the expanding transcript panel, the leading transcript toggle, and the idle close affordance overlay
- move transport rendering, badge state, replay scrubber gating, status text, and playback error presentation into `iOS/KeyVox iOS/Views/HomeTabView/TTS/HomeTabView+TTSTransport.swift`
- move preparation presentation state, button title and enablement logic, selected voice resolution, and TTS action handlers into `iOS/KeyVox iOS/Views/HomeTabView/TTS/HomeTabView+TTSPresentation.swift`
- extract the replay scrubber into `iOS/KeyVox iOS/Views/HomeTabView/TTS/TTSReplayScrubber.swift` as its own focused view and make it internally visible so sibling Home TTS files can reference it after the split
- preserve the existing Home TTS behavior while cleaning up concern boundaries so transcript behavior, transport behavior, presentation state, and section layout are easier to reason about and maintain
