+++ iOS/KeyVox iOS/Views/HomeTabView/TTS/HomeTabView+TTSPresentation.swift
@@ -1,0 +1,140 @@
+import SwiftUI
+
+extension HomeTabView {
+    var ttsButtonTitle: String {
+        if pocketTTSModelManager.isSharedModelReady() == false {
+            return "Install"
+        }
+
+        switch pocketTTSModelManager.installState(for: effectiveTTSVoice) {
+        case .notInstalled:
+            return "Install"
+        case .downloading, .installing:
+            return "Installing"
+        case .failed:
+            return "Repair"
+        case .ready:
+            return ttsManager.isActive ? "Stop" : "Speak"
+        }
+    }
+
+    var showsTTSPreparationProgress: Bool {
+        switch ttsManager.state {
+        case .preparing, .generating:
+            return true
+        case .idle, .playing, .finished, .error:
+            return false
+        }
+    }
+
+    var ttsPreparationProgressLabel: String {
+        "\(Int(ttsManager.playbackPreparationProgress * 100))%"
+    }
+
+    var ttsPreparationPercentageText: String? {
+        showsTTSPreparationProgress ? ttsPreparationProgressLabel : nil
+    }
+
+    func syncTTSPreparationPresentation() {
+        showsTTSPreparationSlot = showsTTSPreparationProgress
+        isTTSPreparationVisible = showsTTSPreparationProgress
+    }
+
+    func updateTTSPreparationPresentation() {
+        ttsPreparationCollapseTask?.cancel()
+
+        if showsTTSPreparationProgress {
+            if showsTTSPreparationSlot == false {
+                showsTTSPreparationSlot = true
+            }
+
+            withAnimation(.easeOut(duration: 0.14)) {
+                isTTSPreparationVisible = true
+            }
+            return
+        }
+
+        guard showsTTSPreparationSlot else { return }
+
+        withAnimation(.easeOut(duration: 0.14)) {
+            isTTSPreparationVisible = false
+        }
+
+        ttsPreparationCollapseTask = Task { @MainActor in
+            try? await Task.sleep(nanoseconds: 240_000_000)
+            guard Task.isCancelled == false else { return }
+
+            withAnimation(.easeInOut(duration: 0.52)) {
+                showsTTSPreparationSlot = false
+            }
+        }
+    }
+
+    var isTTSButtonEnabled: Bool {
+        switch pocketTTSModelManager.sharedModelInstallState {
+        case .downloading, .installing:
+            return false
+        case .notInstalled, .failed, .ready:
+            break
+        }
+
+        switch pocketTTSModelManager.installState(for: effectiveTTSVoice) {
+        case .downloading, .installing:
+            return false
+        case .notInstalled, .failed, .ready:
+            return true
+        }
+    }
+
+    var effectiveTTSVoice: AppSettingsStore.TTSVoice {
+        if pocketTTSModelManager.isVoiceReady(settingsStore.ttsVoice) {
+            return settingsStore.ttsVoice
+        }
+
+        return pocketTTSModelManager.installedVoices().first ?? settingsStore.ttsVoice
+    }
+
+    func handlePrimaryTTSAction() {
+        if pocketTTSModelManager.isSharedModelReady() == false {
+            switch pocketTTSModelManager.sharedModelInstallState {
+            case .notInstalled:
+                pocketTTSModelManager.downloadSharedModel()
+            case .failed:
+                pocketTTSModelManager.repairSharedModelIfNeeded()
+            case .downloading, .installing, .ready:
+                break
+            }
+            return
+        }
+
+        switch pocketTTSModelManager.installState(for: effectiveTTSVoice) {
+        case .notInstalled:
+            pocketTTSModelManager.downloadVoice(effectiveTTSVoice)
+        case .failed:
+            pocketTTSModelManager.repairVoiceIfNeeded(effectiveTTSVoice)
+        case .downloading, .installing:
+            break
+        case .ready:
+            appHaptics.light()
+            audioModeCoordinator.handleSpeakClipboardFromApp()
+        }
+    }
+
+    func handleSecondaryTTSAction() {
+        if ttsManager.isPlaybackPaused {
+            audioModeCoordinator.handleResumeTTS()
+            return
+        }
+
+        if ttsManager.state == .playing {
+            audioModeCoordinator.handlePauseTTS()
+            return
+        }
+
+        audioModeCoordinator.handleReplayLastTTS()
+    }
+
+    func handleReplayScrub(_ progress: Double) {
+        ttsManager.seekReplay(toProgress: progress)
+    }
+}
